Updated to Symphony Framework 3.2.4.
Added the group field as a key on the product file.

Added the ability to search for products in the product view list.  The search allows you to match records on the primary product code key or the group key.  You can select to force a complete key match or partial match.
Ensured that the Symphony.Conductor.Statis.ThreadSafe.AnchorToUI() method is called to anchor the main UI thread.  This ensures that Symphony can correctly marshal back to the UI thread when signalling property changes.

Changed the ProductViewerViewModel to use the AsyncObservableCollection to ensure that any property change notifications are raised on the correct UI thread.